Fix up k8 registry to resolve unit tests
The logic for the registry is pretty special due to namespace checks and using the common logic in the RunnerFilter doesn't work because namespace filters throw off the "have I specified any includes/excludes?" logic. So, I've tried to make the minimum changes possible to meet the main requirements, even if it is a little different than other registries.