Enable checks for terraform module calls (#481)
* Add a context parser for modules.

This is derived from the resource_context_parser.py Like oyu refer to a
module with module.<name> the path is ["module", <name>].

* Add classes for module checks

Also derived from the existing code, I added a base class for writing
checks for modules. It automatically registers the check to a registry
for module checks, as already known from resources etc.

* Add module registry to the terraform runner

* Add tests for module check registry

* Add a module into a test file to test if this causes problems

* Add a test to validate that the runner runs the checks for modules

Co-authored-by: Jonas Eckert <j.eckert@ostfalia.de>