Implement a system to allow old signatures to still be valid
For adding new arguments to checks, we needed a mechanism to add
parameters without breaking code, that doesn't expect them. With this
decorator meta class combination, it is possible to define alternative
signatures. They can be used to call the old signature with the new one
by providing the implementation for the wrapping.