Restructure config creation
Now __init__ takes each possible argument. For args (and later for file)
there is a static function, that calls __init__. It is done this way, to
allow easy programmatic creation of configs.

Otherwise it is not easy to create config instances. They are required
for testing and when checkov is called via python code. For testing you
could use some code that interfaces the old args parameter.