HCL1 transformation attempts and parsing support
This adds the ability to read HCL1 files if the env variable `CKV_ENABLE_HCL1` is set to `yes` (or `1` or `true`). When a terraform file can't be read with the HCL2 parser, the HCL(1) parser is attempted. If that works, there is a transformation process that takes place to convert the data to appear as though it came from the HCL2 parser. This process is by no means perfect (there's a bit of art mixed in with science), but it might be good enough for most circumstances.

Fixes: #681