CKV_AZURE_116: Fix Azure Policies check location
Per the azurerm_kubernetes_cluster documentation, this attribute is set
in the `addon_profile` block; it is not a field at the top level of the
resource definition. This would have resulted in false failures for
resources that did properly define this attribute and false passed for
resources that specified it improperly.

https://registry.terraform.io/providers/hashicorp/azurerm/latest/docs/resources/kubernetes_cluster#azure_policy