Check for wildcards inside of lists
My company just found a case of a dev using a wildcard inside of an AWS principal list to evade this rule.  Since the `principal_value` in this code could be either a string or a list, checking for membership as well as equality would prevent the circumnavigation of the rule.