fix(cloudformation): KMSKeyWildCardPrincipal modification - Check for wildcards inside of lists (#4590)
* Check for wildcards inside of lists

My company just found a case of a dev using a wildcard inside of an AWS principal list to evade this rule.  Since the `principal_value` in this code could be either a string or a list, checking for membership as well as equality would prevent the circumnavigation of the rule.

* type check

Make sure the value given is a list first.  Still might be able to exploit using a !Join of empty string with a wildcard though like `!Join [ '', ['', '*' ] ]`, but intrinsics would have to be handled as a whole prior to checks being run.

* Add test resource

Add wildcard inside of an AWS principal list

* Update fail count with new resource added