platform(general): Fix policy.name to use the spaces as specified on CLI. (#6296)
When adding the flag policy.name with a valid value of a particular policy, only that policy should run and return in the result. Instead all the policies are returned (and hence run). This is fixed by handling the spaces in the policy names and leading and trailing whitespaces appropriately.