fix(general): base_runner: Properly escape excluded directories that begin with '.'
In case the repo has been checked out to a path that inadvertently
matches one of the excluded directory patterns (i.e. .git), we should
check for it, and escape the pattern via ``re.escape()`` instead of just
escaping the leading '.' in '.terraform'.

Otherwise, this causes some tests to fail if the repo is checked out to
``~/git/checkov``, as this would match the '.git' excluded directory by
virtue of the leading '.' not being escaped.