feat(terraform_plan): support scanning deleted resources
Previously, it was not possible to check a resource's configuration when
the plan indicated the resource would be deleted entirely. This prevented
custom checks from enforcing conventions around deletions.

We now process pure delete actions from resource_changes. The "before"
state is made available as the resource configuration, with __change_actions__
set to ["delete"].